Basic GUI operations added.
Buttons are disabled on the main GUI screen.
You must go to the menu bar and create a new file/open a file before anything can be done.
File type to be opened is specified as an *.elv file which stands for "Event Log Viewer File."
Once a file has been created or opened, if you try to either close the window, open a different file, or create a new file, then a popup will come up asking whether or not you want to save the current project.