Export improvements
Improvements:
* Add html export type
* Add comments to export window
* Refactor export window for usability/consistency
Issues:
* For some reason, the position of the check mark in the file type
drop-down is preserved when the export window is closed/opened. The
variable is correctly set to the default, but the check-mark remains on
the old value until it is changed. I can't figure out how to fix it, and
it's only a visual error, so I'm going to let it slide for now.